feat: move util to saner place
