+++ src/pages/ThorChainLP/components/AddLiquidity/AddLiquidityInput.tsx
@@ -66,12 +66,9 @@ import {
 } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
 import { assertUnreachable, isSome, isToken } from 'lib/utils'
 import { getSupportedEvmChainIds } from 'lib/utils/evm'
-import { getThorchainFromAddress } from 'lib/utils/thorchain'
+import { getThorchainFromAddress, getThorchainTransactionType } from 'lib/utils/thorchain'
 import { THOR_PRECISION, THORCHAIN_POOL_MODULE_ADDRESS } from 'lib/utils/thorchain/constants'
-import {
-  estimateAddThorchainLiquidityPosition,
-  getThorchainTransactionType,
-} from 'lib/utils/thorchain/lp'
+import { estimateAddThorchainLiquidityPosition } from 'lib/utils/thorchain/lp'
 import { AsymSide, type LpConfirmedDepositQuote } from 'lib/utils/thorchain/lp/types'
 import { depositWithExpiry } from 'lib/utils/thorchain/routerCalldata'
 import { useGetEstimatedFeesQuery } from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
